Document parts of Playground CLI
