+++ packages/playground/cli/src/download.ts
@@ -5,14 +5,28 @@ import path, { basename } from 'path';
 
 export const CACHE_FOLDER = path.join(os.homedir(), '.wordpress-playground');
 
-async function downloadTo(
-	url: string,
-	destinationPath: string,
+// @TODO: Support HTTP cache, invalidate the local file if the remote file has changed
+export async function cachedDownload(
+	remoteUrl: string,
+	cacheKey: string,
 	monitor: EmscriptenDownloadMonitor
 ) {
-	const response = await monitor.monitorFetch(fetch(url));
+	const artifactPath = path.join(CACHE_FOLDER, cacheKey);
+	if (!fs.existsSync(artifactPath)) {
+		fs.ensureDirSync(CACHE_FOLDER);
+		await downloadTo(remoteUrl, artifactPath, monitor);
+	}
+	return readAsFile(artifactPath);
+}
+
+async function downloadTo(
+	remoteUrl: string,
+	localPath: string,
+	monitor: EmscriptenDownloadMonitor
+) {
+	const response = await monitor.monitorFetch(fetch(remoteUrl));
 	const reader = response.body!.getReader();
-	const writer = fs.createWriteStream(destinationPath);
+	const writer = fs.createWriteStream(localPath);
 	while (true) {
 		const { done, value } = await reader.read();
 		if (done) {
@@ -22,25 +36,11 @@ async function downloadTo(
 	}
 }
 
-// @TODO: Support HTTP cache, invalidate the local file if the remote file has changed
-export async function cachedDownload(
-	url: string,
-	artifactName: string,
-	monitor: EmscriptenDownloadMonitor
-) {
-	const artifactPath = path.join(CACHE_FOLDER, artifactName);
-	if (!fs.existsSync(artifactPath)) {
-		fs.ensureDirSync(CACHE_FOLDER);
-		await downloadTo(url, artifactPath, monitor);
-	}
-	return readAsFile(artifactPath);
-}
-
 export function readAsFile(path: string, fileName?: string): File {
 	return new File([fs.readFileSync(path)], fileName ?? basename(path));
 }
 
-export async function getWordPressVersionUrlAndName(version = 'latest') {
+export async function resolveWPRelease(version = 'latest') {
 	if (version === 'trunk' || version === 'nightly') {
 		return {
 			url: 'https://wordpress.org/nightly-builds/wordpress-latest.zip',
